Refactor tests for just response validation
